+++ example/flutter_what3words_with_google_maps/lib/services/grid_section_lines.dart
@@ -1,8 +1,6 @@
 // NOTE
 // this test grid contained hard-coded 20 polylines to not consume the API
 
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:syncfusion_flutter_maps/maps.dart';