+++ example/grid_section_example.dart
@@ -6,7 +6,8 @@ void main() async {
 
   // Create and execute a request to obtain a grid section within the provided bounding box (sw and ne corners)
   var gridSection = await api
-      .gridSection(Coordinates(51.515900, -0.212517), Coordinates(51.527649, -0.191746))
+      .gridSection(
+          Coordinates(51.515900, -0.212517), Coordinates(51.527649, -0.191746))
       .execute();
 
   if (gridSection.isSuccessful()) {