+++ lib/src/request/autosuggest_options.dart
@@ -12,6 +12,7 @@ class AutosuggestOptions {
   String? _inputType;
   String? _language;
   String? _preferLand;
+  String? _locale;
 
   String? get nResults => _nResults;
   String? get focus => _focus;
@@ -23,6 +24,7 @@ class AutosuggestOptions {
   String? get inputType => _inputType;
   String? get language => _language;
   String? get preferLand => _preferLand;
+  String? get locale => _locale;
 
   ///Set the number of AutoSuggest results to return. A maximum of 100 results can be specified, if a number greater than this is requested,
   ///this will be truncated to the maximum. The default is 3
@@ -94,9 +96,10 @@ class AutosuggestOptions {
   ///[southWest] Southwest coordinates of the bounding box
   ///[northEast] Northeast coordinates of the bounding box
   ///Returns an [AutosuggestOptions] instance suitable to allow more query customisation
-  AutosuggestOptions setClipToBoundingBox(Coordinates southWest, Coordinates northEast) {
+  AutosuggestOptions setClipToBoundingBox(
+      Coordinates southWest, Coordinates northEast) {
     _clipToBoundingBox =
-    '${southWest.lat},${southWest.lng},${northEast.lat},${northEast.lng}';
+        '${southWest.lat},${southWest.lng},${northEast.lat},${northEast.lng}';
     return this;
   }
 
@@ -125,6 +128,12 @@ class AutosuggestOptions {
     return this;
   }
 
+  /// Sets the locale for the AutoSuggest request
+  AutosuggestOptions setLocale(String locale) {
+    _locale = locale; // Setter for locale
+    return this;
+  }
+
   ///For power users, used to specify voice input mode. Can be [AutosuggestInputType.TEXT] (default), [AutosuggestInputType.VOCON_HYBRID],
   ///[AutosuggestInputType.NMDP_ASR]. See voice recognition section within the developer docs for more details https://docs.what3words.com/api/v3/#voice.
   ///
@@ -140,5 +149,4 @@ class AutosuggestOptions {
     _preferLand = '$preferLand';
     return this;
   }
-
 }