+++ lib/src/request/convert_to_3wa_request.dart
@@ -3,7 +3,6 @@ import 'package:what3words/src/response/response.dart';
 import '../response/location.dart';
 import 'coordinate.dart';
 import 'abstract_builder.dart';
-import '../service/what3words_v3.dart';
 import 'request.dart';
 
 class ConvertTo3WARequest extends Request<Location> {
@@ -16,19 +15,19 @@ class ConvertTo3WARequest extends Request<Location> {
         super(builder.api);
 
   Future<Response<Location>> execute() async {
-    return await super.call(
-        api.what3words().convertTo3wa, [coordinates, language]);
+    return await super
+        .call(api.what3words().convertTo3wa, [coordinates, language]);
   }
 }
 
 /// Builder for `convert-to-3wa` API requests
-class ConvertTo3WARequestBuilder extends AbstractBuilder<Future<Response<Location>>> {
+class ConvertTo3WARequestBuilder
+    extends AbstractBuilder<Future<Response<Location>>> {
   final String _coordinates;
   String? _language;
 
-  ConvertTo3WARequestBuilder(What3WordsV3 api, Coordinates coordinates)
-      : _coordinates = '${coordinates.lat}, ${coordinates.lng}',
-        super(api);
+  ConvertTo3WARequestBuilder(super.api, Coordinates coordinates)
+      : _coordinates = '${coordinates.lat}, ${coordinates.lng}';
 
   ///Set the language to return the 3 word address in
   ///