+++ lib/src/service/json_to_type_converter.dart
@@ -24,7 +24,8 @@ class JsonToTypeConverter extends JsonConverter {
 
     if (jsonMap is List) {
       return jsonMap
-          .map((item) => jsonParser!(item as Map<String, dynamic>) as InnerType?)
+          .map(
+              (item) => jsonParser!(item as Map<String, dynamic>) as InnerType?)
           .toList() as T;
     }