+++ lib/src/service/what3words_v3.dart
@@ -9,10 +9,11 @@ import '../request/convert_to_coordinates_request.dart';
 import '../request/coordinate.dart';
 import '../request/grid_section_request.dart';
 import 'what3words_service.dart';
+import 'package:what3words/src/response/isvalid3wa_response.dart';
 
 ///Instances of the What3WordsV3 class provide access to Version 3 of the what3words API.
 class What3WordsV3 {
-  final _defaultEndpoint = 'https://api.what3words.com/v3';
+  final _defaultEndpoint = "api.what3words.com/v3";
   late What3WordsV3Service _service;
 
   ///Get a new API manager instance.
@@ -42,6 +43,64 @@ class What3WordsV3 {
   void _setupHttpClient(
       String apiKey, String endpoint, Map<String, String>? headers) {
     _service = What3WordsV3Service.create(apiKey, endpoint, headers);
+  }
+
+  /// Check if a string is a possible what3words address using regex.
+  ///
+  /// [text] The string to check.
+  /// Returns true if it's a match against the regex, false otherwise.
+  bool isPossible3wa(String text) {
+    var regex = RegExp(r"^\/{0,}(?:[^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}[.｡。･・︒។։။۔።।][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}[.｡。･・︒។։။۔።।][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}|[^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}([\u0020\u00A0][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]+){1,3}[.｡。･・︒។։။۔።।][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}([\u0020\u00A0][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]+){1,3}[.｡。･・︒។։။۔።।][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}([\u0020\u00A0][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]+){1,3})$");
+    return regex.hasMatch(text);
+  }
+
+  /// Check if a string is a possible what3words address with different separators.
+  ///
+  /// [text] The string to check.
+  /// Returns true if it's a match against the regex, false otherwise.
+  bool didYouMean3wa(String text) {
+    var regex = RegExp(r"[^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}[.｡。･・︒។։။۔።। ,\\\\^_/+'&:;|　-]{1,2}[^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}[.｡。･・︒។։။۔።। ,\\\\^_/+'&:;|　-]{1,2}[^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r';:£§º©®\s]{1,}');
+    return regex.hasMatch(text);
+  }
+
+  /// Find all possible what3words addresses in a string.
+  ///
+  /// [text] The string to search.
+  /// Returns a list of possible what3words addresses.
+  List<String> findPossible3wa(String text) {
+    var regex = RegExp(r"[^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}[.｡。･・︒។։။۔።।][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r";:£§º©®\s]{1,}[.｡。･・︒។։။۔።।][^0-9`~!@#$%^&*()+\-_=[{\]}\\|'<,.>?\/"
+        '"'
+        r';:£§º©®\s]{1,}');
+    var matches = regex.allMatches(text);
+    return matches.map((match) => match.group(0) ?? "").toList();
   }
 
   ///Convert a 3 word address to a latitude and longitude. It also returns country, the bounds of the grid square,
@@ -83,7 +142,8 @@ class What3WordsV3 {
   ///bounding-box that crosses the anti-meridian, use longitude greater than 180.
   ///
   ///returns a [GridSectionRequestBuilder] instance suitable for invoking a `grid-section` API request
-  GridSectionRequestBuilder gridSection(Coordinates southwest, Coordinates northeast) {
+  GridSectionRequestBuilder gridSection(
+      Coordinates southwest, Coordinates northeast) {
     return GridSectionRequestBuilder(this, southwest, northeast);
   }
 
@@ -96,7 +156,8 @@ class What3WordsV3 {
   ///[options] The autosuggest options and clippings, check available options here https://docs.what3words.com/api/v3/#autosuggest
   ///
   ///returns a [AutosuggestRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestRequestBuilder autosuggest(String input, {AutosuggestOptions? options}) {
+  AutosuggestRequestBuilder autosuggest(String input,
+      {AutosuggestOptions? options}) {
     return AutosuggestRequestBuilder(this, input, options);
   }
 
@@ -109,7 +170,9 @@ class What3WordsV3 {
   ///[options] The autosuggest options, check available options here https://docs.what3words.com/api/v3/#autosuggest
   ///
   ///returns a [AutosuggestWithCoordinatesRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestWithCoordinatesRequestBuilder autosuggestWithCoordinates(String input, {AutosuggestOptions? options}) {
+  AutosuggestWithCoordinatesRequestBuilder autosuggestWithCoordinates(
+      String input,
+      {AutosuggestOptions? options}) {
     return AutosuggestWithCoordinatesRequestBuilder(this, input, options);
   }
 
@@ -122,8 +185,68 @@ class What3WordsV3 {
   ///[rank] The rank of the selected suggestion.
   ///[options] The autosuggest options used on the autosuggest or autosuggest-with-coordinates call.
   ///returns a [AutosuggestWithCoordinatesRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestSelectionRequestBuilder autosuggestSelection(String rawInput, String sourceApi, String words, int rank, {AutosuggestOptions? options}) {
-    return AutosuggestSelectionRequestBuilder(this, rawInput, sourceApi, words, rank, options);
+  AutosuggestSelectionRequestBuilder autosuggestSelection(
+      String rawInput, String sourceApi, String words, int rank,
+      {AutosuggestOptions? options}) {
+    return AutosuggestSelectionRequestBuilder(
+        this, rawInput, sourceApi, words, rank, options);
+  }
+
+  /// This function checks whether a given what3words address is valid.
+  ///
+  /// The function first verifies if the input string is a possible
+  /// what3words address format by using the `isPossible3wa` method.
+  /// If the format is invalid, it returns an error indicating
+  /// `BAD_WORDS`.
+  ///
+  /// If the input format is correct, the function then calls the
+  /// `autosuggest` API with a limit of one result. The function
+  /// evaluates the response from the API:
+  /// - If the API request fails, it checks if the failure is due to
+  /// an invalid API key (`INVALID_KEY`) or some other error, and
+  /// returns the appropriate error response.
+  /// - If the request is successful and the first suggestion matches
+  /// the input address, it returns a success response indicating
+  /// that the address is valid.
+  ///
+  /// If no valid match is found, the function returns a `BAD_WORDS`
+  /// error, indicating that the address is not recognized as a valid
+  /// what3words address.
+  ///
+  /// Returns:
+  /// - `IsValid3waResponse`: A custom response object indicating
+  /// whether the address is valid and if any errors occurred.
+  ///
+  /// Example:
+  /// ```dart
+  /// var response = await api.isValid3wa("filled.count.soap");
+  /// if (response.isSuccessful() && response.isValid == true) {
+  ///   print("Valid address");
+  /// } else {
+  ///   print("Invalid address or error: ${response.error?.message}");
+  /// }
+  /// ```
+  Future<IsValid3waResponse> isValid3wa(String words) async {
+    if (!isPossible3wa(words)) {
+      return IsValid3waResponse.success(false);
+    }
+
+    var autosuggestResult =
+        await autosuggest(words, options: AutosuggestOptions().setNResults(1))
+            .execute();
+
+    if (!autosuggestResult.isSuccessful()) {
+      return IsValid3waResponse.error(autosuggestResult.error()!);
+    }
+
+    for (var suggestion in autosuggestResult.data()?.suggestions ?? []) {
+      if (suggestion.words.replaceAll("/", "").toLowerCase() ==
+          words.replaceAll("/", "").toLowerCase()) {
+        return IsValid3waResponse.success(true);
+      }
+    }
+
+    return IsValid3waResponse.success(false);
   }
 
   What3WordsV3Service what3words() {