+++ test/autosuggest_focus_test.dart
@@ -1,7 +1,5 @@
 import 'dart:io' show Platform;
-
 import 'package:test/test.dart';
-import 'package:what3words/src/request/autosuggest_options.dart';
 import 'package:what3words/what3words.dart';
 
 void main() {
@@ -11,10 +9,10 @@ void main() {
     var options = AutosuggestOptions().setFocus(Coordinates(51.2, 0.2));
     var autosuggest =
         await api.autosuggest('index.home.ra', options: options).execute();
-    expect(autosuggest.isSuccessful(), true);
 
     var suggestions = autosuggest.data()!.suggestions;
-
+    // print(suggestions);
+    expect(autosuggest.isSuccessful(), true);
     var found = false;
     for (var s in suggestions) {
       if (s.words == 'index.home.raft') {
@@ -65,9 +63,8 @@ void main() {
   test('testFocusSmallLongitude', () async {
     var options = AutosuggestOptions().setFocus(Coordinates(51.2, -360));
 
-    var autosuggest = await api
-        .autosuggest('index.home.ra', options: options)
-        .execute();
+    var autosuggest =
+        await api.autosuggest('index.home.ra', options: options).execute();
     expect(autosuggest.isSuccessful(), true);
 
     var suggestions = autosuggest.data()!.suggestions;
@@ -76,6 +73,32 @@ void main() {
     for (var s in suggestions) {
       if (s.words == 'index.home.raft') {
         found = true;
+      }
+    }
+    expect(found, true);
+  });
+
+  // New test case to test locale parameter
+  test('testLocaleWithAutoSuggest', () async {
+    var options = AutosuggestOptions()
+        .setFocus(Coordinates(51.2, 0.2))
+        .setLanguage('oo'); // Use a language that has locales
+
+    var autosuggest = await api
+        .autosuggest('бонтон.популаран.гус', options: options)
+        .execute();
+
+    expect(autosuggest.isSuccessful(), true);
+
+    var suggestions = autosuggest.data()!.suggestions;
+
+    var found = false;
+    for (var s in suggestions) {
+      if (s.words == 'бонтон.популаран.гусле') {
+        found = true;
+        // Verify the locale was returned correctly
+        expect(s.language, 'oo');
+        expect(s.locale, 'oo_cy');
       }
     }
     expect(found, true);